+++ src/main/java/com/ustudents/examples/scenes/EcsExample2.java
@@ -57,7 +57,7 @@ public class EcsExample2 extends Scene {
     }
 
     @Override
-    public void update(double dt) {
+    public void update(float dt) {
 
     }
